Solved issue with duplicate code. Split MovementOutputData into itself and the new EnemyOutputData. MovementOutputData exclusively contains info on the directions the player can move in and is now used by multiple presenters. EnemyOutputData contains info on the enemy that the MovementInteractor sometimes needs to generate. Deleted other OutputData classes that were replaced by MovementOutputData
