+++ src/use_case/movement/MovementOutputBoundary.java
@@ -4,6 +4,6 @@ public interface MovementOutputBoundary {
 
     void prepareEmptyRoomView(MovementOutputData movementOutputData);
     void prepareStairsView(MovementOutputData movementOutputData);
-    void prepareTurnSelectView(MovementOutputData movementOutputData);
+    void prepareTurnSelectView(EnemyOutputData enemyOutputData);
     void prepareItemCollectionView(MovementOutputData movementOutputData);
 }
