+++ src/use_case/stairs/NextFloorOutputBoundary.java
@@ -1,5 +1,7 @@
 package use_case.stairs;
 
+import use_case.movement.MovementOutputData;
+
 public interface NextFloorOutputBoundary {
-    void prepareStartingRoomView(NextFloorOutputData nextFloorOutputData);
+    void prepareStartingRoomView(MovementOutputData MovementOutputData);
 }
