--- src/use_case/stairs/NextFloorOutputData.java
@@ -1,14 +1,0 @@
-package use_case.stairs;
-
-public class NextFloorOutputData {
-
-    private String directions;
-
-    public NextFloorOutputData(String directions){
-        this.directions = directions;
-    }
-
-    public String getDirections() {
-        return directions;
-    }
-}
