+++ src/use_case/stairs/StayOnFloorOutputBoundary.java
@@ -1,6 +1,8 @@
 package use_case.stairs;
 
+import use_case.movement.MovementOutputData;
+
 public interface StayOnFloorOutputBoundary {
 
-    void prepareEmptyRoomView(StayOnFloorOutputData stayOnFloorOutputData);
+    void prepareEmptyRoomView(MovementOutputData movementOutputData);
 }
