--- src/use_case/stairs/StayOnFloorOutputData.java
@@ -1,14 +1,0 @@
-package use_case.stairs;
-
-public class StayOnFloorOutputData {
-
-    private String directions;
-
-    public StayOnFloorOutputData(String directions){
-        this.directions = directions;
-    }
-
-    public String getDirections() {
-        return directions;
-    }
-}
